+++ packaging.R
@@ -80,6 +80,10 @@ use_github_action("check-standard")
 use_github_action("test-coverage")
 use_github_action("pkgdown")
 
+# Revdep
+use_revdep()
+
+
 #=============================================================================
 # Finish package building (can use fresh session)
 #=============================================================================
@@ -99,6 +103,9 @@ install(upgrade = FALSE)
 if (FALSE) {
   check_win_devel()
   check_rhub()
+  
+  # Takes long
+  revdepcheck::revdep_check(num_workers = 4)
 
   # Wait until above checks are passed without relevant notes/warnings
   # then submit to CRAN