+++ NEWS.md
@@ -6,7 +6,7 @@
 
 ## Small changes
 
-- Adding $(n x K)$ matrix of predictions to the output, where $n$ is the number of rows in the explainer data `X`, and $K$ is the dimension of a single prediction (usually 1).
+- Adding $(n \times K)$ matrix of predictions to the output, where $n$ is the number of rows in the explainer data `X`, and $K$ is the dimension of a single prediction (usually 1).
 
 # kernelshap 0.3.4