+++ NEWS.md
@@ -10,6 +10,11 @@
 - Removed the `ks_extract()` function. It was designed to extract objects like the matrix `S` of SHAP values from the resulting "kernelshap" object `x`. We feel that the standard extraction options (`x$S`, `x[["S"]]`, or `getElement(x, "S")`) are sufficient.
 - Adding $(n \times K)$ matrix of predictions to the output, where $n$ is the number of rows in the explainer data `X`, and $K$ is the dimension of a single prediction (usually 1).
 
+## Other changes
+
+- More unit tests
+
+
 # kernelshap 0.3.4
 
 ## Documentation