+++ NEWS.md
@@ -1,3 +1,9 @@
+# kernelshap 0.3.6
+
+## Maintenance
+
+- Replace term "exact SHAP" by "exact Kernel SHAP" in README and help.
+
 # kernelshap 0.3.5
 
 ## Maintenance